requirements.txt
tweepy
kafka-python
streamlit
boto3
wordcloud
matplotlib
transformers
pyspark
fastapi
uvicorn